Decimal to flaot

+++ server/migrations/20200614044629-create-income.js
@@ -12,7 +12,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       amount: {
-        type: Sequelize.DECIMAL(10,2),
+        type: Sequelize.FLOAT
       },
       description: {
         type: Sequelize.STRING
